refactor PlanEntity in subFiles
