+++ lib/plan/plan_itinerary_tabs.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
+import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
 import 'package:trufi_core/l10n/trufi_localization.dart';
 
 import '../widgets/vertical_swipe_detector.dart';
