+++ lib/services/plan_request/request_manager.dart
@@ -1,5 +1,5 @@
 import 'package:async/async.dart';
-import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
+import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
 import 'package:trufi_core/trufi_models.dart';
 
 abstract class RequestManager {
